requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
Click==7.0
elasticsearch==6.3.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
pytz==2018.9
PyYAML==3.13
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
